Centralizada funcion connectDB en el fichero server/db/index.js
