+++ .travis.yml
@@ -4,6 +4,6 @@ node_js:
 services:
     - docker
 before_install:
-    - docker run -e MYSQL_DATABASE=license_management -e MYSQL_ROOT_PASSWORD=admin -d -p 3310:3306 mysql:5.7 
+    - docker run -e MYSQL_DATABASE=licenses_management_test -e MYSQL_ROOT_PASSWORD=root -d -p 3310:3306 mysql:5.7 
 script:
-    - BD_PORT=3310 BD_NAME=license_management DB_PASSWORD=admin npm run test
+    - DB_PORT=3310 DB_PASSWORD=root npm run test
