+++ test/models/model_licenses.test.js
@@ -1,6 +1,7 @@
 const assert = require('assert');
 //Incializamos sequelize
 const sequelize = require('../../server/db').sequelize;
+const connectDB = require('../../server/db').connectDB;
 //Cargamos los modelos
 const models = require('../../server/models')(sequelize);
 
@@ -10,7 +11,12 @@ describe('Model Licenses', function() {
 	
   beforeEach(function(done){
 	  //Sincronizamos la base de datos
-	  sequelize.sync({force:true}).then(() => done());
+	  sequelize.sync({force:true}).then( () => {
+	  		return done();
+		}, (err) => {
+			console.log("Error connecting DB, retrying...");
+			setTimeout(connectDB, 5000);
+		})
   });
 
   it('Should create a empty models', function(done) {
