+++ test/relations/relations_user_org.test.js
@@ -2,6 +2,8 @@ const assert = require('assert');
 const sequelize_fixtures = require('sequelize-fixtures');
 //Incializamos sequelize
 const sequelize = require('../../server/db').sequelize;
+const connectDB = require('../../server/db').connectDB;
+
 //Cargamos los modelos
 const models = require('../../server/models')(sequelize);
 
@@ -10,12 +12,15 @@ const models = require('../../server/models')(sequelize);
 describe('Relations between User-Organizations', function() {
   
   beforeEach(function(done){
-   //Sincronizamos la base de datos
-      sequelize.sync({force:true}).then(function(){
+    //Sincronizamos la base de datos
+    sequelize.sync({force:true}).then( () => {
       //Cargamos los datos necesarios
       sequelize_fixtures.loadFile('test/fixtures/fixtures.json', models)
             .then(() => done())
-      })
+    }, (err) => {
+      console.log("Error connecting DB, retrying...");
+      setTimeout(connectDB, 5000);
+    })
   });
