+++ test/routes/change_profile_server.test.js
@@ -6,6 +6,7 @@ const sequelize_fixtures = require('sequelize-fixtures');
 const passwordHash = require('password-hash');
 //Incializamos sequelize
 const sequelize = require('../../server/db').sequelize;
+const connectDB = require('../../server/db').connectDB;
 
 //Cargamos los modelos
 const models = require('../../server/models')(sequelize);
@@ -20,11 +21,14 @@ describe('Profile Test', function() {
 
   //Before each test we clean databse and load fixtures file.
   beforeEach(function(done){
-    //Sincronizamos la base de datos
-    sequelize.sync({force:true}).then(function(){
+       //Sincronizamos la base de datos
+    sequelize.sync({force:true}).then( () => {
       //Cargamos los datos necesarios
       sequelize_fixtures.loadFile('test/fixtures/fixtures.json', models)
             .then(() => done())
+    }, (err) => {
+      console.log("Error connecting DB, retrying...");
+      setTimeout(connectDB, 5000);
     })
   });
  it('Should return a 200 Ok message. Login ok and current password correct', function(done) {
